UserManagement::deleteUser: check existence of personal file area top folder before deleting it, re #11500
git-svn-id: svn://develop.studip.de/studip/trunk@61677 c5b85792-5220-0410-ad23-ed5b6b8ad999

+++ lib/classes/UserManagement.class.php
@@ -979,8 +979,10 @@ class UserManagement
 
         if ($delete_personal_documents) {
             $user_folder = Folder::findTopFolder($this->user->id);
-            $this->msg .= 'info§' . _('Persönlicher Dateibereich gelöscht.') . '§';
-            $user_folder->delete();
+            if ($user_folder) {
+                $this->msg .= 'info§' . _('Persönlicher Dateibereich gelöscht.') . '§';
+                $user_folder->delete();
+            }
             NotificationCenter::postNotification('UserDataDidRemove', $this->user_data['auth_user_md5.user_id'], 'personal_documents');
         }
